updating test with accessibility identifier and readme added
