+++ MoviesUITests/Tests/MoviesLaunchPerformanceTests.swift
@@ -1,0 +1,29 @@
+//
+//  MoviesUITests.swift
+//  MoviesUITests
+//
+//  Created by Ahmed Abdurrahman on 3/25/20.
+//  Copyright © 2020 VNGRS. All rights reserved.
+//
+
+
+import Foundation
+import XCTest
+
+class MoviesLaunchPerformanceTests: BaseTest {
+    let homePage = HomePage()
+    
+    override func setUp() {
+        super.setUp()
+        waitForPageLoading(element: homePage.popular)
+    }
+
+    func testLaunchPerformance() {
+        if #available(macOS 10.15, iOS 13.0, tvOS 13.0, *) {
+            // This measures how long it takes to launch your application.
+            measure(metrics: [XCTOSSignpostMetric.applicationLaunch]) {
+                XCUIApplication().launch()
+            }
+        }
+    }
+}
